refactor: Refactor
Moved a bunch of code around to try and prevent rendering of layout hook before its ready -- does not seem to fix the issue
