+++ src/components/Dashboard/Dashboard.css
@@ -1,18 +1,5 @@
 @import '@folio/stripes-components/lib/variables.css';
 
-/**
- * Pane Content
- */
-
-.dashboardContent {
-  height: 100%;
-  overflow-y: scroll;
-  overflow: auto;
-  outline: 0;
-  padding: var(--gutter-static);
-  position: relative
-}
-
 .widgetsContainer {
   display: flex;
   height: 100%;
