+++ src/components/DashboardContainer/DashboardContainer.css
@@ -16,3 +16,16 @@
   position: absolute;
   will-change: transform;
 }
+
+/**
+ * Pane Content
+ */
+
+.dashboardContent {
+  height: 100%;
+  overflow-y: scroll;
+  overflow: auto;
+  outline: 0;
+  padding: var(--gutter-static);
+  position: relative
+}
