+++ src/components/DashboardContainer/DashboardContainer.js
@@ -3,15 +3,25 @@
  * along with the actions menu and dashboard tab groups.
  * This will ALSO be used to render the actions menu for the "no dashboards" splash screen
  */
-
+import { useCallback, useState } from 'react';
+import { FormattedMessage } from 'react-intl';
 import PropTypes from 'prop-types';
+
+import {
+  ConfirmationModal,
+  Loading
+} from '@folio/stripes/components';
 
 import Header from './Header';
 import css from './DashboardContainer.css';
-import Dashboard from '../Dashboard';
+import Dashboard, { NoWidgets } from '../Dashboard';
+
+import { ErrorModal } from '../ErrorComponents';
+import DashboardAccessInfo from '../DashboardAccessInfo/DashboardAccessInfo';
 
 const DashboardContainer = ({
   dashboard,
+  dashboardLoading,
   dashboards,
   onCreateDashboard,
   onCreateWidget,
@@ -24,6 +34,65 @@ const DashboardContainer = ({
   onWidgetEdit,
   widgets
 }) => {
+  // Handle delete through a delete confirmation modal rather than directly
+  const [showDeleteConfirmationModal, setShowDeleteConfirmationModal] = useState(false);
+
+  // Keep track of which widget we're deleting--necessary because this is the dashboard level
+  const [widgetToDelete, setWidgetToDelete] = useState({});
+
+  // This stores the CANVAS-LEVEL error state, ready to display in the modal
+  const [errorState, setErrorState] = useState({
+    errorMessage: null,
+    errorModalOpen: false,
+    errorStack: null
+  });
+
+  // This takes an error and a stacktrace to pass to the modal, and opens it
+  const handleError = useCallback((err, stack) => {
+    setErrorState({
+      ...errorState,
+      errorMessage: err,
+      errorModalOpen: true,
+      errorStack: stack
+    });
+  }, [errorState]);
+
+  const handleHideModal = () => {
+    setErrorState({
+      ...errorState,
+      errorModalOpen: false
+    });
+  };
+
+  const setupConfirmationModal = (widgetId, widgetName) => {
+    // Hijack the onDelete function to show confirmation modal instead at this level
+    setShowDeleteConfirmationModal(true);
+    setWidgetToDelete({ name: widgetName, id: widgetId });
+  };
+
+  const dashboardContents = () => {
+    if (dashboardLoading) {
+      return (
+        <Loading />
+      );
+    }
+
+    if (!widgets?.length) {
+      return (
+        <NoWidgets />
+      );
+    }
+
+    return (
+      <Dashboard
+        handleError={handleError}
+        onWidgetEdit={onWidgetEdit}
+        setupConfirmationModal={setupConfirmationModal}
+        widgets={widgets}
+      />
+    );
+  };
+
   return (
     <>
       <div className={css.dashboardContainer}>
@@ -39,11 +108,40 @@ const DashboardContainer = ({
           onReorder={widgets?.length > 1 ? onReorder : null}
           onUserAccess={onUserAccess}
         />
-        <Dashboard
-          dashboard={dashboard}
-          onWidgetDelete={onWidgetDelete}
-          onWidgetEdit={onWidgetEdit}
-          widgets={widgets}
+        <div className={css.dashboardContent}>
+          <DashboardAccessInfo dashId={dashboard.id} />
+          {dashboardContents()}
+        </div>
+        <ConfirmationModal
+          buttonStyle="danger"
+          confirmLabel={<FormattedMessage id="ui-dashboard.delete" />}
+          data-test-delete-confirmation-modal
+          heading={<FormattedMessage id="ui-dashboard.dashboard.deleteWidget" />}
+          id="delete-agreement-confirmation"
+          message={
+            <FormattedMessage
+              id="ui-dashboard.dashboard.deleteWidgetConfirmMessage"
+              values={{ name: widgetToDelete.name }}
+            />
+          }
+          onCancel={() => {
+            setShowDeleteConfirmationModal(false);
+            setWidgetToDelete({});
+          }}
+          onConfirm={() => {
+            onWidgetDelete(widgetToDelete.id);
+            setShowDeleteConfirmationModal(false);
+            setWidgetToDelete({});
+          }}
+          open={showDeleteConfirmationModal}
+        />
+        <ErrorModal
+          handlers={{
+            onHideModal: handleHideModal
+          }}
+          message={errorState.errorMessage}
+          modalOpen={errorState.errorModalOpen}
+          stack={errorState.errorStack}
         />
       </div>
     </>
@@ -56,6 +154,7 @@ DashboardContainer.propTypes = {
     id: PropTypes.string.isRequired,
     name: PropTypes.string.isRequired
   }).isRequired,
+  dashboardLoading: PropTypes.bool,
   dashboards: PropTypes.arrayOf(PropTypes.shape({
     dashboard: PropTypes.shape({
       id: PropTypes.string.isRequired,
